split classes into separate files, implement add_random_households
