+++ Assignment1/question1.py
@@ -1,7 +1,9 @@
 import pandas as pd
 import numpy as np
 from matplotlib import pyplot as plt
-from classes import Household, Neighborhood, Appliance
+from appliance import Appliance
+from household import Household
+from neighborhood import Neighborhood
 from helper_functions import get_appliances
 
 
@@ -12,7 +14,7 @@ from helper_functions import get_appliances
     - The washing machine should be done when the person is awake so that it can be put in the dryer
 """
 
-appliance_dict = get_appliances()
+appliance_dict = get_appliances(output_dict=True)
 ev = appliance_dict["EV"]
 washing_machine = appliance_dict["Laundry machine"]
 dishwasher = appliance_dict["Dishwasher"]
