+++ Assignment1/question2.py
@@ -2,7 +2,8 @@ import pandas as pd
 import numpy as np
 from matplotlib import pyplot as plt
 from helper_functions import get_appliances
-from classes import Household, Neighborhood
+from household import Household
+from neighborhood import Neighborhood
 
 def generate_pricing_data():
     pricing = np.zeros(24)
@@ -13,17 +14,19 @@ def generate_pricing_data():
 
 pricing = np.load('data/rt_pricing.npy')
 
-lonely_house = Household("lonely house")
-print(get_appliances(filter_shiftable=0))
+# lonely_house = Household("lonely house")
+# print(get_appliances(filter_shiftable=0))
 
-lonely_house.add_appliances(get_appliances(filter_shiftable=0)) 
-lonely_house.add_appliances(get_appliances(filter_shiftable=1))
-#app_to_be_saved = get_random()
-print(lonely_house)
+# lonely_house.add_appliances(get_appliances(filter_shiftable=0)) 
+# lonely_house.add_appliances(get_appliances(filter_shiftable=1))
+# print(lonely_house)
 #lonely_house.add_appliances(get_random_optional_shiftable())
-random_neighborhood = Neighborhood("another lonely", [lonely_house], pricing="RTP")
+
+random_neighborhood = Neighborhood("another lonely", pricing="RTP")
+random_neighborhood.add_random_households(1)
 c, u, l, A_eq, b_eq, A_ub, b_ub = random_neighborhood.get_linprog_input()
 #print(A_eq, b_eq)
 random_neighborhood.optimize()
 
-#print(random_neighborhood.get_schedule().reshape(-1,24))
+print(random_neighborhood.houses[0])
+print(random_neighborhood.get_schedule().reshape(-1,24))
